Update bower.json
Main in bower.json should only be the primary files, not all the files. If you use a tools such as wiredep, it will include the main files, and that will end with all the rxjs files included, which is not what you want.

Also, ignore should ignore any files not needed in production. 

http://bower.io/docs/creating-packages/